python-split-list/requirements.txt
Pillow==9.3.0
more-itertools==9.0.0
numpy==1.23.5
